♻️ refactor : store type도 따로 분리  #158
